test cases working
