+++ endpoints/decorators/metrics.go
@@ -30,8 +30,6 @@ func (w *writerWithStatus) Header() http.Header {
 
 func MonitorHttp(handler httprouter.Handle, me *metrics.Metrics, method string) httprouter.Handle {
 	return httprouter.Handle(func(resp http.ResponseWriter, req *http.Request, params httprouter.Params) {
-		//entry.Request.Mark(1)
-		//me.Add(fmt.Sprintf("%s.current_url.request_count", method), nil, "")
 		logMetrics(me, method, "add", nil)
 		wrapper := writerWithStatus{
 			delegate: resp,
@@ -42,16 +40,10 @@ func MonitorHttp(handler httprouter.Handle, me *metrics.Metrics, method string) 
 		respCode := wrapper.statusCode
 		// If the calling function never calls WriterHeader explicitly, Go auto-fills it with a 200
 		if respCode == 0 || respCode >= 200 && respCode < 300 {
-			//entry.Duration.UpdateSince(start)
-			//me.Add(fmt.Sprintf("%s.current_url.request_duration", method), &start, "")
 			logMetrics(me, method, "", &start)
 		} else if respCode >= 400 && respCode < 500 {
-			//entry.BadRequest.Mark(1)
-			//me.Add(fmt.Sprintf("%s.current_url.bad_request_count", method), nil, "")
 			logMetrics(me, method, "bad_request", nil)
 		} else {
-			//entry.Errors.Mark(1)
-			//me.Add(fmt.Sprintf("%s.current_url.error_count", method), nil, "")
 			logMetrics(me, method, "error", nil)
 		}
 	})
