+++ metrics/prometheus.go
@@ -47,9 +47,7 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 	requestSizeBuckts := []float64{0.001, 0.002, 0.005, 0.01, 0.025, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 1} // TODO: tweak
 	registry := prometheus.NewRegistry()
 	promMetrics := &PrometheusMetrics{
-		//Registry        *prometheus.Registry
 		Registry: registry,
-		//Puts            *PrometheusRequestStatusMetric
 		Puts: &PrometheusRequestStatusMetric{
 			Duration: newHistogram(cfg, registry,
 				"puts.current_url.request_duration", //modify according to InfluxDB name
@@ -62,7 +60,6 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 				[]string{"status"}, // CounterVec labels --> "status": "add", "error", or "bad_request"
 			), //{"puts.current_url.error_count", "puts.current_url.bad_request_count", "puts.current_url.request_count", "gets.current_url.error_count", "gets.current_url.bad_request_count", "gets.current_url.request_count", "puts.backend.error_count", "puts.backend.bad_request_count", "puts.backend.json_request_count", "puts.backend.xml_request_count","puts.backend.defines_ttl", "puts.backend.unknown_request_count", "gets.backend.error_count", "gets.backend.bad_request_count", "gets.backend.request_count"}
 		},
-		//Gets            *PrometheusRequestStatusMetric
 		Gets: &PrometheusRequestStatusMetric{
 			Duration: newHistogram(cfg, registry,
 				"gets.current_url.request_duration",
@@ -75,14 +72,12 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 				[]string{"status"}, // CounterVec labels --> "status": "add", "error", or "bad_request"
 			), //{"gets.current_url.error_count", "gets.current_url.bad_request_count", "gets.current_url.request_count"}
 		},
-		//PutsBackend     *PrometheusRequestStatusMetricByFormat
 		PutsBackend: &PrometheusRequestStatusMetricByFormat{
 			Duration: newHistogram(cfg, registry,
 				"puts.backend.request_duration",
 				"Duration in seconds Prebid Cache takes to process backend put requests.",
 				cacheWriteTimeBuckts,
 			),
-			//PutBackendRequests *prometheus.CounterVec
 			PutBackendRequests: newCounterVecWithLabels(cfg, registry,
 				"puts.backend",
 				"Count of total requests to Prebid Cache labeled by format, status and whether or not it comes with TTL",
@@ -95,7 +90,6 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 				requestSizeBuckts,
 			),
 		},
-		//GetsBackend     *PrometheusRequestStatusMetric
 		GetsBackend: &PrometheusRequestStatusMetric{
 			Duration: newHistogram(cfg, registry,
 				"gets.backend.request_duration",
@@ -109,7 +103,6 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 			), //{"gets.backend.error_count", "gets.backend.bad_request_count", "gets.backend.request_count"}
 
 		},
-		//Connections     *PrometheusConnectionMetrics
 		Connections: &PrometheusConnectionMetrics{
 			ConnectionsOpened: newGaugeMetric(cfg, registry,
 				"connections",
@@ -121,8 +114,6 @@ func CreatePrometheusMetrics(cfg config.PrometheusMetrics) *PrometheusMetrics {
 				[]string{"connection_error"},
 			), // "connection_error" = {"accept", "close"}
 		},
-
-		//ExtraTTLSeconds *prometheus.HistogramVec
 		ExtraTTL: &PrometheusExtraTTLMetrics{
 			ExtraTTLSeconds: newHistogram(cfg, registry,
 				"extra_ttl_seconds",
@@ -204,62 +195,8 @@ func newHistogramVector(cfg config.PrometheusMetrics, registry *prometheus.Regis
 /**************************************************
  *	DEPECRATED Functions to record metrics
  **************************************************/
-// Export begins sending metrics to the configured database.
-// This method blocks indefinitely, so it should probably be run in a goroutine.
 func (m PrometheusMetrics) Export(cfg config.Metrics) {
-	//logrus.Infof("Metrics will be exported to Prometheus with host=%s, db=%s, username=%s", cfg.Influx.Host, cfg.Influx.Database, cfg.Influx.Username)
-	//influxdb.InfluxDB(
-	//	m.Registry,          // metrics registry
-	//	time.Second*10,      // interval
-	//	cfg.Influx.Host,     // the InfluxDB url
-	//	cfg.Influx.Database, // your InfluxDB database
-	//	cfg.Influx.Username, // your InfluxDB user
-	//	cfg.Influx.Password, // your InfluxDB password
-	//)
-	return
 }
-
-//func (m PrometheusMetrics) Increment(metricName string, start *time.Time, value string) {
-//	metricNameTokens := strings.Split(metricName, ".")
-//
-//	if len(metricNameTokens) == 2 && metricNameTokens[0] == "connections" {
-//		switch metricNameTokens[1] {
-//		case "close_errors":
-//			fallthrough
-//		case "accept_errors":
-//			m.ConnectionErrorMetrics.With(prometheus.Labels{
-//				metricNameTokens[0]: metricNameTokens[1], // { "connections.accept_errors", "connections.close_errors"}
-//			}).Inc()
-//		case "active_incoming":
-//			m.ActiveConnections.Inc() //{ "connections.active_incoming"}
-//		}
-//	} else if len(metricNameTokens) == 3 {
-//		label := fmt.Sprintf("%s.%s", metricNameTokens[0], metricNameTokens[1])
-//		if metricNameTokens[0] == "gets" || metricNameTokens[0] == "puts" {
-//			if metricNameTokens[2] == "request_duration" {
-//				m.RequestDurationMetrics.With(prometheus.Labels{"method": label, "result": metricNameTokens[2]}).Observe(time.Since(*start).Seconds())
-//				// {"puts.current_url.request_duration", "gets.current_url.request_duration", "puts.backend.request_duration", "gets.backend.request_duration"}
-//			} else if metricNameTokens[2] == "request_size_bytes" {
-//				m.RequestSyzeBytes.With(prometheus.Labels{
-//					"method": fmt.Sprintf("%s.%s", label, metricNameTokens[2]), // {"puts.current_url.request_duration", "gets.current_url.request_duration", "puts.backend.request_duration", "gets.backend.request_duration"}
-//				}).Observe(float64(len(value)))
-//			} else {
-//				m.MethodToEndpointMetrics.With(prometheus.Labels{
-//					"method": label, "count_type": metricNameTokens[2], //{"puts.current_url.error_count", "puts.current_url.bad_request_count", "puts.current_url.request_count", "gets.current_url.error_count", "gets.current_url.bad_request_count", "gets.current_url.request_count", "puts.backend.error_count", "puts.backend.bad_request_count", "puts.backend.json_request_count", "puts.backend.xml_request_count","puts.backend.defines_ttl", "puts.backend.unknown_request_count", "gets.backend.error_count", "gets.backend.bad_request_count", "gets.backend.request_count"}
-//				}).Inc()
-//			}
-//		}
-//	}
-//}
-//
-//func (m PrometheusMetrics) Decrement(metricName string) {
-//	switch metricName {
-//	case "connections.active_incoming":
-//		m.ActiveConnections.Dec()
-//	default:
-//		//error
-//	}
-//}
 
 /**************************************************
  *	NEW Functions to record metrics
@@ -299,7 +236,7 @@ func (m *PrometheusMetrics) RecordExtraTTLSeconds(value float64) {
 }
 
 /**************************************************
- *	NEW Auxiliary functions to record metrics
+ *	Auxiliary functions to record metrics
  **************************************************/
 func incCounterInVector(counter *prometheus.CounterVec, label string, status string, labels []string) {
 	for _, aLabel := range labels {
