+++ server/listener.go
@@ -23,13 +23,9 @@ type monitorableConnection struct {
 func (l *monitorableConnection) Close() error {
 	err := l.Conn.Close()
 	if err == nil {
-		//l.metrics.ActiveConnections.Dec(1)
-		//l.metrics.Substract("connections.active_incoming")
 		l.metrics.RecConnectionMetrics("substract")
 	} else {
 		log.Errorf("Error closing connection: %v", err)
-		//l.metrics.ConnectionCloseErrors.Mark(1)
-		//l.metrics.Add("connections.close_errors", nil, "")
 		l.metrics.RecConnectionMetrics("close")
 	}
 	return err
@@ -39,13 +35,9 @@ func (ln *monitorableListener) Accept() (c net.Conn, err error) {
 	tc, err := ln.Listener.Accept()
 	if err != nil {
 		log.Errorf("Error accepting connection: %v", err)
-		//ln.metrics.ConnectionAcceptErrors.Mark(1)
-		//ln.metrics.Add("connections.accept_errors", nil, "")
 		ln.metrics.RecConnectionMetrics("accept")
 		return tc, err
 	}
-	//ln.metrics.ActiveConnections.Inc(1)
-	//ln.metrics.Add("connections.active_incoming", nil, "")
 	ln.metrics.RecConnectionMetrics("add")
 	return &monitorableConnection{
 		tc,
